+++ deal/src/main/java/com/deal/model/dto/CreditDTO.java
@@ -1,4 +1,4 @@
-package com.deal.model.DTO;
+package com.deal.model.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
