--- deal/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.application.name=deal
-spring.datasource.url=jdbc:postgresql://localhost:5432/credit_conveyor_db
-spring.datasource.username=postgres
-spring.datasource.password=password
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.hibernate.ddl-auto = create-drop
